+++ packages/extended-text-entry/src/index.js
@@ -40,6 +40,12 @@ export default class RootExtendedTextEntry extends HTMLElement {
     this._session = null;
   }
 
+  setLangAttribute() {
+    const language = this._model && typeof this._model.language ? this._model.language : '';
+    const lang = language ? language.slice(0, 2) : 'en';
+    this.setAttribute('lang', lang);
+  }
+
   set model(m) {
     this._model = m;
     this.dispatchEvent(new ModelSetEvent(this.tagName.toLowerCase(), false, !!this._model));
@@ -96,6 +102,8 @@ export default class RootExtendedTextEntry extends HTMLElement {
         onAnnotationsChange: this.annotationsChange.bind(this),
         onCommentChange: this.commentChange.bind(this),
       });
+
+      this.setLangAttribute();
 
       ReactDOM.render(elem, this, () => {
         renderMath(this);