+++ packages/inline-dropdown/src/index.js
@@ -28,6 +28,12 @@ export default class RootInlineDropdown extends HTMLElement {
     return this._session;
   }
 
+  setLangAttribute() {
+    const language = this._model && typeof this._model.language ? this._model.language : '';
+    const lang = language ? language.slice(0, 2) : 'en';
+    this.setAttribute('lang', lang);
+  }
+
   _render = () => {
     if (this._model && this._session) {
       let elem = React.createElement(InlineDropdown, {
@@ -44,6 +50,8 @@ export default class RootInlineDropdown extends HTMLElement {
         language:  this._model.language,
         onChange: this.changeSession,
       });
+
+      this.setLangAttribute();
 
       ReactDOM.render(elem, this, () => {
         renderMath(this);