+++ packages/math-templated/controller/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.16.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/math-templated-controller@2.15.0...@pie-element/math-templated-controller@2.16.0) (2024-06-25)
+
+
+### Bug Fixes
+
+* fixed getResponseCorrectness return object not being updated correctly ([14d557f](https://github.com/pie-framework/pie-elements/commit/14d557fc17b166a66c5083a30a60229fd57cc15d))
+
+
+### Features
+
+* **math-templated:** Added unit tests for controller ([cb41fbd](https://github.com/pie-framework/pie-elements/commit/cb41fbd35d665f51b615a54167bbd9b874aee17a))
+
+
+
+
+
+# [2.15.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/math-templated-controller@2.14.0...@pie-element/math-templated-controller@2.15.0) (2024-06-24)
+
+
+### Features
+
+* update pie-lib/pie-toolbox PD-3868, PD-3771, PD-3772, PD-3775 ([4475c65](https://github.com/pie-framework/pie-elements/commit/4475c658be5489a0f66be27af24e01c03b32b294))
+
+
+
+
+
 # [2.14.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/math-templated-controller@2.13.0...@pie-element/math-templated-controller@2.14.0) (2024-06-24)